--- anime-list-api/app/tests/sampel.test.js
@@ -1,6 +1,0 @@
-describe('Sample Test', () => {
-    it('should test that true === true', () => {
-      expect(true).toBe(true)
-    });
-});
-
